rename scribe.elasticsearch -> hub.elastic_sync
