+++ hub/elastic_sync/fast_ar_trending.py
