--- hub/elasticsearch/__init__.py
@@ -1,2 +1,0 @@
-from hub.elasticsearch.search import SearchIndex
-from hub.elasticsearch.notifier_protocol import ElasticNotifierClientProtocol
